+++ src/driver/motors.h
@@ -6,8 +6,8 @@ class Motors {
 public:
 	explicit Motors(Motor m1, Motor m2);
   void init();
-	void deploy(int speedPercent);
-	void retract(int speedPercent);
+	void deploy(uint8_t speedPercent);
+	void retract(uint8_t speedPercent);
   void stop();
 	Motor motor1;
 	Motor motor2;
